Reworking 360 Degree PDFs
These new 360˚ PDFs should be able to direct the police car through traffic. 

The only file you need Ricky is 'pdf_function.py'. The function you call is shown on line 21 and it takes 4 arguments. The first is the list of all coords of civilians wrt to the police car. Then it takes the police car coords, just default these to [0,0]. The third argument is the heading of the car, so if the road is 15 degrees to the LEFT of the police car, the car is at heading +15˚, basically clockwise from the road. That's the convention I'm using. And the final argument is is the angle you are testing for, i.e. finding the probability for that proposed angle. 

Within the function, on line 21, it's actual code in it's separate file allows for more tweaking of parameters during the simulation, and if things aren't working then lemme know and I'll tell you how to update different things about the code.